[lesson] Refactor lesson fetch
