+++ src/pages/api/lesson/[_id].ts
@@ -1,6 +1,4 @@
 import { lessonGetOneHandler } from "@handlers/lesson/get.handler";
-import { bookmarkAddHandler } from "@handlers/user/bookmark/add.handler";
-import { bookmarkDeleteHandler } from "@handlers/user/bookmark/delete.handler";
 import routerMiddleware from "@middlewares/router.middleware";
 import { ApiResponse } from "@typing/api-response.interface";
 import { NextApiHandler, NextApiRequest, NextApiResponse } from "next";
@@ -15,8 +13,6 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
   const { _id } = req.query as { _id: string };
   const handlers: Record<string, NextApiHandler<ApiResponse>> = {
     GET: lessonGetOneHandler(_id),
-    POST: bookmarkAddHandler(_id),
-    DELETE: bookmarkDeleteHandler(_id),
     // add here handlers for other methods
   };
