+++ src/server/services/lessons.service.ts
@@ -125,3 +125,159 @@ export const updateBookmarkCounter = async (
     { $inc: { bookmarkCount: amount } }
   );
 };
+
+// A [key, value] tuple
+type QueryEntry = [string, string | string[]];
+
+/**
+ * Checks that the given value is actually a single string instead of an array of strings.
+ *
+ * @param {string | string[]} value Value read from request.
+ * @param {string} message Error message to send back.
+ * @throws {Error} if `value` is an array.
+ */
+const forbidMultipleValues = (
+  value: string | string[],
+  message?: string
+): void => {
+  if (Array.isArray(value)) {
+    throw new Error(message || "les valeurs multiples sont interdites");
+  }
+};
+
+/**
+ * Make sure the given value is an array of strings.
+ *
+ * @param {string | string[]} value
+ * @return {string[]}
+ */
+const toArray = (value: string | string[]): string[] =>
+  Array.isArray(value) ? value : [value];
+
+/**
+ * Converts the query to an object containing filters to use with the database.
+ *
+ * @param {Record<string, string|string[]>} rawQuery The query from the request
+ * @return {Filter<ILessonDB>} Filters
+ */
+export const getFiltersFromQuery = (
+  rawQuery: Record<string, string | string[]>
+): Filter<ILessonDB> => {
+  const query: QueryEntry[] = Object.entries(rawQuery);
+  const filters: Filter<ILessonDB> = {};
+
+  for (const [key, value] of query) {
+    switch (key) {
+      // Search by author id (multiple values are treated as "or")
+      case "author":
+        filters.authorId = { $in: toArray(value) };
+        break;
+
+      // Text search (in title and subtitle)
+      case "title":
+      case "subtitle":
+        forbidMultipleValues(
+          value,
+          "On ne peut chercher qu'un titre/sous-titre à la fois"
+        );
+        filters.$text = {
+          $search: value as string,
+          $caseSensitive: false,
+        };
+        break;
+
+      // Search by draft status
+      case "isDraft":
+        filters.isDraft = JSON.parse(value as string);
+        break;
+
+      // Search by category id (multiple values are treated as "and")
+      case "category":
+        filters.categoryIds = {
+          $all: toArray(value),
+        };
+        break;
+
+      // Search by tag id (multiple values are treated as "and")
+      case "tag":
+        filters.tagIds = {
+          $all: toArray(value),
+        };
+        break;
+
+      // Search by creation date
+      case "creationDateBefore":
+        forbidMultipleValues(
+          value,
+          "On ne peut avoir qu'une seule limite de fin pour la date de création"
+        );
+        filters.creationDate = {
+          ...filters.creationDate,
+          $lte: new Date(value as string),
+        };
+        break;
+      case "creationDateAfter":
+        forbidMultipleValues(
+          value,
+          "On ne peut avoir qu'une seule limite de début pour la date de création"
+        );
+        filters.creationDate = {
+          ...filters.creationDate,
+          $gte: new Date(value as string),
+        };
+        break;
+
+      // Search by lastModified date
+      case "lastModifiedDateBefore":
+        forbidMultipleValues(
+          value,
+          "On ne peut avoir qu'une seule limite de fin pour la date de modification"
+        );
+        filters.lastModifiedDate = {
+          ...filters.lastModifiedDate,
+          $lte: new Date(value as string),
+        };
+        break;
+      case "lastModifiedDateAfter":
+        forbidMultipleValues(
+          value,
+          "On ne peut avoir qu'une seule limite de début pour la date de modification"
+        );
+        filters.lastModifiedDate = {
+          ...filters.lastModifiedDate,
+          $gte: new Date(value as string),
+        };
+        break;
+
+      // Search by publication date
+      case "publicationDateBefore":
+        forbidMultipleValues(
+          value,
+          "On ne peut avoir qu'une seule limite de fin pour la date de publication"
+        );
+        filters.publicationDate = {
+          ...filters.publicationDate,
+          $lte: new Date(value as string),
+        };
+        break;
+      case "publicationDateAfter":
+        forbidMultipleValues(
+          value,
+          "On ne peut avoir qu'une seule limite de début pour la date de publication"
+        );
+        filters.publicationDate = {
+          ...filters.publicationDate,
+          $gte: new Date(value as string),
+        };
+        break;
+      case "bookmarks":
+        // TODO restrict the lessons to those in the current user's bookmarks
+        break;
+
+      default:
+        throw new Error(`Filtre inconnu : '${key}'`);
+    }
+  }
+
+  return filters;
+};
